fix: vkui token class only mode=full
