+++ src/components/ConfigProvider/ConfigProvider.tsx
@@ -1,8 +1,5 @@
 import * as React from 'react';
-import {
-  AppearanceProvider,
-  generateVKUITokensClassName,
-} from '../AppearanceProvider/AppearanceProvider';
+import { AppearanceProvider } from '../AppearanceProvider/AppearanceProvider';
 import {
   ConfigProviderContext,
   ConfigProviderContextInterface,
@@ -11,8 +8,6 @@ import {
 import { useObjectMemo } from '../../hooks/useObjectMemo';
 import { useAutoDetectAppearance } from '../../hooks/useAutoDetectAppearance';
 import { noop } from '@vkontakte/vkjs';
-import { useIsomorphicLayoutEffect } from '../../lib/useIsomorphicLayoutEffect';
-import { useDOM } from '../../lib/dom';
 
 export interface ConfigProviderProps extends Partial<ConfigProviderContextInterface> {
   onDetectAppearanceByBridge?: () => void;
@@ -40,20 +35,6 @@ export const ConfigProvider = (props: ConfigProviderProps) => {
   };
 
   const appearance = useAutoDetectAppearance(appearanceProp, onDetectAppearanceByBridge);
-
-  const { document } = useDOM();
-
-  useIsomorphicLayoutEffect(() => {
-    const VKUITokensClassName = generateVKUITokensClassName(platform, appearance);
-
-    // eslint-disable-next-line no-restricted-properties
-    document!.body.classList.add(VKUITokensClassName);
-
-    return () => {
-      // eslint-disable-next-line no-restricted-properties
-      document!.body.classList.remove(VKUITokensClassName);
-    };
-  }, [platform, appearance]);
 
   const configContext = useObjectMemo({
     webviewType,
